move merge and validate into interface
