+++ lang/python/airr/__init__.py
@@ -1,7 +1,7 @@
 """
 Reference library for AIRR schema for Ig/TCR rearrangements
 """
-from airr.interface import read, create, derive, load, write
+from airr.interface import read, create, derive, load, write, merge, validate
 
 # versioneer-generated
 from ._version import get_versions
