+++ lang/python/airr/tools.py
@@ -19,73 +19,11 @@ AIRR tools and utilities
 
 # System imports
 import argparse
-import os
 import sys
-from collections import OrderedDict
-from itertools import chain
 
 # Local imports
 from airr import __version__
-from airr.io import RearrangementReader, RearrangementWriter
-
-def merge(out_file, airr_files, drop=False, debug=False):
-    """
-    Merge one or more AIRR rearrangements files
-
-    Arguments:
-      out_file (str): output file name.
-      airr_files (list): list of input files to merge.
-      drop (bool): drop flag. If True then drop fields that do not exist in all input
-                   files, otherwise combine fields from all input files.
-      debug (bool): debug flag. If True print debugging information to standard error.
-
-    Returns:
-      bool: True if files were successfully merged, otherwise False.
-    """
-    try:
-        # gather fields from input files
-        readers = [RearrangementReader(open(f, 'r'), debug=debug) for f in airr_files]
-        field_list = [x.fields for x in readers]
-        if drop:
-            field_set = set.intersection(*map(set, field_list))
-        else:
-            field_set = set.union(*map(set, field_list))
-        field_order = OrderedDict([(f, None) for f in chain(*field_list)])
-        out_fields = [f for f in field_order if f in field_set]
-
-        out_handle = open(out_file, 'w')
-        writer = RearrangementWriter(out_handle, fields=out_fields, debug=debug)
-
-        for reader in readers:
-            for rec in reader:
-                writer.write(rec)
-
-        out_handle.close()
-    except:
-        return False
-
-    return True
-
-
-def validate(airr_files, debug=False):
-    """
-    Validates one or more AIRR rearrangements files
-
-    Arguments:
-      airr_files (list): list of input files to validate.
-      debug (bool): debug flag. If True print debugging information to standard error.
-
-    Returns:
-      boolean: True if all files passed validation, otherwise False
-    """
-    valid = True
-    for file in airr_files:
-        print('Validating: %s' % os.path.basename(file))
-        reader = RearrangementReader(open(file, 'r'))
-        valid &= reader.validate()
-
-    return valid
-
+from airr.interface import merge, validate
 
 def define_args():
     """
